backend/sfx/: simplify SFX request creation - parse request body params from querystring and get XML for POST body
